Configure tasks that need "embulkPluginExtension" only when they are needed
"EmbulkPluginExtension" has been always validated when running Gradle so that
the "jar" and "gem" tasks can use the extension values.

The validation has been sometimes annoying, however. For example, just running
"./gradlew wrapper --gradle-version=5.6" does not work if "build.gradle" does
not contain "embulkPlugin { ... }" configured properly.

This commit makes the validation lazy so that it works only when the "jar" or
"gem" tasks are executed, by externaizing the initializer methods into separate
tasks, and getting them executed by the "dependsOn" mechanism of Gradle.

It adds task groups and descriptions, and bumps up to 0.2.6-SNAPSHOT along with
the change.
