+++ build.gradle
@@ -10,7 +10,7 @@ plugins {
 // They want Gradle plugins to be published under the "gradle.plugin" prefix for some security reasons.
 // https://plugins.gradle.org/docs/publish-plugin
 group = "org.embulk"
-version = "0.2.5-SNAPSHOT"
+version = "0.2.6-SNAPSHOT"
 description = "A Gradle plugin to build and publish Embulk plugins"
 
 repositories {
