+++ src/main/java/org/embulk/gradle/embulk_plugins/WarnIfRuntimeHasCompileOnlyDependencies.java
@@ -1,0 +1,91 @@
+/*
+ * Copyright 2019 The Embulk project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.embulk.gradle.embulk_plugins;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+import javax.inject.Inject;
+import org.gradle.api.DefaultTask;
+import org.gradle.api.Project;
+import org.gradle.api.artifacts.Configuration;
+import org.gradle.api.artifacts.ResolvedDependency;
+import org.gradle.api.logging.Logger;
+import org.gradle.api.tasks.TaskAction;
+
+class WarnIfRuntimeHasCompileOnlyDependencies extends DefaultTask {
+    @Inject
+    public WarnIfRuntimeHasCompileOnlyDependencies(final Configuration alternativeRuntimeConfiguration) {
+        super();
+        this.alternativeRuntimeConfiguration = alternativeRuntimeConfiguration;
+    }
+
+    @TaskAction
+    public void warn() {
+        final Project project = this.getProject();
+
+        final Configuration compileOnlyConfiguration = project.getConfigurations().getByName("compileOnly");
+
+        final Map<String, ResolvedDependency> compileOnlyDependencies =
+                collectAllDependencies(compileOnlyConfiguration);
+        final Map<String, ResolvedDependency> alternativeRuntimeDependencies =
+                collectAllDependencies(this.alternativeRuntimeConfiguration);
+
+        final Set<String> intersects = new HashSet<>();
+        intersects.addAll(alternativeRuntimeDependencies.keySet());
+        intersects.retainAll(compileOnlyDependencies.keySet());
+        if (!intersects.isEmpty()) {
+            final Logger logger = project.getLogger();
+
+            // Logging in the "error" loglevel to show the severity, but not to fail with GradleException.
+            logger.error(
+                    "============================================ WARNING ============================================\n"
+                    + "Following \"runtime\" dependencies are included also in \"compileOnly\" dependencies.\n"
+                    + "\n"
+                    + intersects.stream().map(key -> {
+                        final ResolvedDependency dependency = alternativeRuntimeDependencies.get(key);
+                        return "  \"" + dependency.getModule().toString() + "\"\n";
+                    }).collect(Collectors.joining(""))
+                    + "\n"
+                    + "  \"compileOnly\" dependencies are used to represent Embulk's core to be \"provided\" at runtime.\n"
+                    + "  They should be excluded from \"compile\" or \"runtime\" dependencies like the example below.\n"
+                    + "\n"
+                    + "  dependencies {\n"
+                    + "    compile(\"org.glassfish.jersey.core:jersey-client:2.25.1\") {\n"
+                    + "      exclude group: \"javax.inject\", module: \"javax.inject\"\n"
+                    + "    }\n"
+                    + "  }\n"
+                    + "=================================================================================================");
+        }
+    }
+
+    private static Map<String, ResolvedDependency> collectAllDependencies(final Configuration configuration) {
+        final HashMap<String, ResolvedDependency> allDependencies = new HashMap<>();
+
+        final Set<ResolvedDependency> firstLevel = configuration.getResolvedConfiguration().getFirstLevelModuleDependencies();
+        for (final ResolvedDependency dependency : firstLevel) {
+            Util.recurseAllDependencies(dependency, allDependencies);
+        }
+
+        return Collections.unmodifiableMap(allDependencies);
+    }
+
+    private final Configuration alternativeRuntimeConfiguration;
+}
